Starting to tinker with Dagger, Repositories, and ViewModels
