+++ app/src/main/java/re/spitfy/ctftime/utils/FirestoreRequestListener.kt
@@ -1,0 +1,8 @@
+package re.spitfy.ctftime.utils
+
+interface FirestoreRequestListener<T> {
+
+    fun onSuccess(response: T)
+
+    fun onFailure(error: Throwable)
+}
